refactor(estoque): modulariza insumos e movimentacoes
