+++ AGENTS.md
@@ -70,3 +70,4 @@ Data | Autor | Descrição
 2025-06-09 | CODEX | Validação completa do módulo de fornecedores com tipagem e organização.
 
 2025-06-10 | CODEX | Módulo de produtos estruturado com serviço e tipagens. Página integra DataTable.
+2025-06-10 | CODEX | Módulo de estoque validado com tipagem e organização.
