+++ MODULOS/ESTOQUE.md
@@ -1,0 +1,19 @@
+# Módulo de Estoque
+
+Campos principais dos Insumos:
+- id
+- name
+- sku
+- description
+- group_id
+- location_id
+- unit_of_measurement_id
+- quantity
+- min_quantity
+- max_quantity
+- cost_price
+- is_active
+- created_at
+- updated_at
+
+Fluxo básico: cadastro de insumos, controle de movimentações de entrada, saída e ajustes, além de alertas de estoque mínimo e gerenciamento de localizações e grupos.
