+++ src/app/(dashboard)/estoque/grupos/_components/GroupForm.tsx
@@ -17,7 +17,7 @@ import { Input } from "@/components/ui/input";
 import { Textarea } from "@/components/ui/textarea";
 import { createRecord, updateRecord } from "@/lib/data-hooks";
 import { toast } from "sonner";
-import type { Group } from "./GroupColumns";
+import type { GrupoDeInsumo } from "@/modules/estoque/estoque.types";
 
 // Define Zod schema based on the 'stock_groups' table
 const groupFormSchema = z.object({
@@ -29,7 +29,7 @@ const groupFormSchema = z.object({
 type GroupFormValues = z.infer<typeof groupFormSchema>;
 
 interface GroupFormProps {
-  initialData?: Partial<Group>; // For editing
+  initialData?: Partial<GrupoDeInsumo>; // For editing
   onSuccess?: () => void; // Callback after successful submission
 }
